+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ## 1.4.3 (2022-08-19)
 
-- Fixed handling of unparseable YAML files in manager methods.
+- Fixed handling of invalid YAML files in manager methods.
 
 ## 1.4.2 (2022-07-29)